==> notebooks/3_Definition_Labels_Ratio.ipynb <==
import requests
import pandas as pd
import numpy as np
import csv
import matplotlib.pyplot as plt
import seaborn as sns
from sklearn.model_selection import train_test_split
from sklearn.preprocessing import StandardScaler
from sklearn.model_selection import GridSearchCV
from sklearn.model_selection import  KFold
import warnings
import re
warnings.filterwarnings('ignore')

movies_db = pd.read_csv("movies_db_2.csv", delimiter=",",engine='python') #Provient de 2_Preprocessing

from sklearn.preprocessing import MinMaxScaler
scaler = MinMaxScaler((-1,1))
#movies_db.drop(index=265,inplace=True)
scaler.fit(movies_db[['vote_average','vote_count',"revenue"]])
data_scaled = scaler.transform(movies_db[['vote_average','vote_count','revenue']])
data_scaled = pd.DataFrame(data_scaled)
data_scaled.columns = ['vote_average','vote_count','revenue']
data_scaled

benefice_ratio=(movies_db["revenue"]-movies_db["budget"])/movies_db["revenue"]
# print(benefice_ratio.describe())
benefice_ratio=pd.DataFrame(benefice_ratio)
scaler = MinMaxScaler((-1,1))
scaler.fit(benefice_ratio)
data_scaled["benefice_ratio"] = scaler.transform(benefice_ratio)
# ratio = 0.3*data_scaled["vote_average"]+0.5*(data_scaled['benefice_ratio'])+0.2*data_scaled["vote_count"]
ratio = (data_scaled["vote_average"]+data_scaled['benefice_ratio']+data_scaled["vote_count"])/3
ratio.describe()

movies_db['ratio'] = ratio
movies_db.sort_values("ratio",ascending=False)

def class_df(row):
  # SUPER HIT
  if row["ratio"] >= 0.5:
    return "SUPER HIT"
  # HIT
  elif 0.25<= row["ratio"] <.5 :
    return "HIT"
  # BOF
  elif 0<= row["ratio"] <0.25 :
    return "BOF"
  # FlOP
  else:
    return "FLOP"

movies_db["target"]=movies_db.apply(lambda row: class_df(row), axis=1)

movies_db.to_csv("movies_db_3.csv",index=False)
