==> notebooks/5_Scaling.ipynb <==
import requests
import pandas as pd
import numpy as np
import csv
import matplotlib.pyplot as plt
import seaborn as sns
from sklearn.model_selection import train_test_split
from sklearn.preprocessing import StandardScaler
from sklearn.model_selection import GridSearchCV
from sklearn.model_selection import  KFold
import warnings
import re
warnings.filterwarnings('ignore')

movies_db = pd.read_csv("movies_db_3.csv", delimiter=",",engine='python') #Provient de 3_Definition_Labels_X

scaler = StandardScaler()
scaler.fit(movies_db.drop(["target","title"],axis=1))
movies_model_data_scaled = scaler.transform(movies_db.drop(["target","title"],axis=1))
movies_model_data_scaled = pd.DataFrame(movies_model_data_scaled)
movies_model_data_scaled.columns = movies_db.drop(["target","title"],axis=1).columns
movies_model_data_scaled["target"] = movies_db["target"]
movies_model_data_scaled["title"] = movies_db["title"]
movies_model_data_scaled

scaler.mean_[[1,3,4,7,8]], scaler.var_[[1,3,4,7,8]]

Export CSV

movies_model_data_scaled.to_csv("movies_db_5.csv",index=False)
