==> notebooks/6_Models.ipynb <==
import requests
import pandas as pd
import numpy as np
import csv
import matplotlib.pyplot as plt
import seaborn as sns
from sklearn.model_selection import train_test_split
from sklearn.preprocessing import StandardScaler
from sklearn.model_selection import GridSearchCV
from sklearn.model_selection import  KFold
import warnings
import re
warnings.filterwarnings('ignore')

movies_model_data_scaled = pd.read_csv("movies_db_5.csv", delimiter=",",engine='python') #Provient de 5_Scaling

movies_model_data_scaled.drop(["id",'vote_average','vote_count','revenue'],axis=1,inplace=True)

movies_model_data_scaled

data_0 = movies_model_data_scaled[movies_model_data_scaled["target"] == 0]
data_1 = movies_model_data_scaled[movies_model_data_scaled["target"] == 1]
data_2 = movies_model_data_scaled[movies_model_data_scaled["target"] == 2]
# data_3 = movies_model_data_scaled[movies_model_data_scaled["target"] == "SUPER HIT"]

x_0 = data_0.drop(["target","title"],axis=1)
x_1 = data_1.drop(["target","title"],axis=1)
x_2 = data_2.drop(["target","title"],axis=1)
# x_3 = data_3.drop(["target","title"],axis=1)

y_0 = data_0["target"]
y_1 = data_1["target"]
y_2 = data_2["target"]
# y_3 = data_3["target"]

x_0_train, x_0_test, y_0_train, y_0_test = train_test_split(x_0, y_0, test_size=0.33)
x_1_train, x_1_test, y_1_train, y_1_test = train_test_split(x_1, y_1, test_size=0.33)
x_2_train, x_2_test, y_2_train, y_2_test = train_test_split(x_2, y_2, test_size=0.33)
# x_3_train, x_3_test, y_3_train, y_3_test = train_test_split(x_3, y_3, test_size=0.33)

x_train = pd.concat([x_0_train,x_1_train,x_2_train])
x_test = pd.concat([x_0_test,x_1_test,x_2_test])
y_train = pd.concat([y_0_train,y_1_train,y_2_train])
y_test = pd.concat([y_0_test,y_1_test,y_2_test])

"""
x_train = x_train.drop(["pop_company", "pop_actors", "pop_producers"],axis=1)
x_test = x_test.drop(["pop_company", "pop_actors", "pop_producers"],axis=1)
"""


from sklearn.neighbors import KNeighborsClassifier
from sklearn.metrics import precision_score
from sklearn.svm import SVC
from sklearn.ensemble import RandomForestClassifier
from sklearn.discriminant_analysis import LinearDiscriminantAnalysis as LDA
from sklearn.discriminant_analysis import QuadraticDiscriminantAnalysis as QDA
import xgboost as xgb
from sklearn.ensemble import GradientBoostingClassifier
from sklearn.ensemble import AdaBoostClassifier
from sklearn.decomposition import PCA
from sklearn.naive_bayes import GaussianNB

### Model sans Grid Search

KNN

knn = KNeighborsClassifier(n_neighbors=4)

knn.fit(x_train,y_train)

y_pred_knn = knn.predict(x_test)
precision_score(y_test, y_pred_knn, average='micro')

LDA

lda = LDA()

lda.fit(x_train,y_train)

y_pred_lda = lda.predict(x_test)
precision_score(y_test, y_pred_lda, average='micro')

QDA

qda = QDA()

qda.fit(x_train,y_train)

y_pred_qda = qda.predict(x_test)
precision_score(y_test, y_pred_qda, average='micro')

Naive Bayes

nb = GaussianNB()

nb.fit(x_train,y_train)

y_pred_nb = nb.predict(x_test)
precision_score(y_test, y_pred_nb, average='micro')

SVC

svc = SVC()

svc.fit(x_train,y_train)

y_pred_svc = svc.predict(x_test)
precision_score(y_test, y_pred_svc, average='micro')

RANDOM FOREST

rfc = RandomForestClassifier()

rfc.fit(x_train,y_train)

y_pred_rfc = rfc.predict(x_test)
precision_score(y_test, y_pred_rfc, average='micro')

XGBOOST

XgBoost = xgb.XGBClassifier()
XgBoost.fit(x_train, y_train)
y_pred_xgb = XgBoost.predict(x_test)
precision_score(y_test, y_pred_xgb, average='micro')

GRADIENT BOOSTING

GradientBoost = GradientBoostingClassifier()
GradientBoost.fit(x_train, y_train)
y_pred_gb = GradientBoost.predict(x_test)
precision_score(y_test, y_pred_gb, average='micro')

ADABOOST

AdaBoost = AdaBoostClassifier()
AdaBoost.fit(x_train, y_train)
y_pred_ab = AdaBoost.predict(x_test)
precision_score(y_test, y_pred_ab, average='micro')

### Model avec Grid Search


LDA

param = {'solver':["svd", "lsqr", "eigen"], 
         'shrinkage':["auto", 0.1, 0.3, 0.8],'n_components': [1,2,3]}
        # , 'learning_decay': [0.5, 0.7, 0.9]
lda_grid = GridSearchCV(LDA(), param, cv= KFold(n_splits=5, random_state=42, shuffle = True), scoring ="accuracy")
lda_grid.fit(x_train, y_train)
y_pred_lda = lda_grid.predict(x_test)
precision_score(y_test, y_pred_lda, average="micro"), lda_grid

QDA

param = {'reg_param':[0, 0.01, 0.1, 0.2, 0.25, 0.5]}

qda_grid = GridSearchCV(QDA(), param, cv= KFold(n_splits=5, random_state=42, shuffle = True), scoring ="accuracy")
qda_grid.fit(x_train, y_train)
y_pred_qda = qda_grid.predict(x_test)
precision_score(y_test, y_pred_qda, average="micro"), qda_grid

KNN

#Hyperparameters definition
param = {"n_neighbors": list(range(1,12)), 
         "weights": ["uniform", "distance"],'metric' : ['minkowski','euclidean','manhattan'],
         "algorithm":["auto", "ball_tree", "kd_tree", "brute"],
         "p":[1,2,3]}
        #"leaf_size" : list(range(1,50)
#GridSearch
knn_grid = GridSearchCV(KNeighborsClassifier(), param, cv= KFold(n_splits=5, random_state=42, shuffle = True),scoring = "accuracy")
knn_grid.fit(x_train, y_train)
y_pred_knn = knn_grid.predict(x_test)
precision_score(y_test, y_pred_knn, average="micro"), knn_grid


SVC

# Hyperparameters
param = {"C": [0.1,1,10,100],
        "kernel":["linear", "poly", "rbf", "sigmoid"],
        'gamma': [0.5, 0.1, 0.001, 0.0001]}
#GridSearch
svc_grid = GridSearchCV(SVC(), param, cv= KFold(n_splits=5, random_state=42, shuffle = True),scoring = "accuracy")
svc_grid.fit(x_train, y_train)
y_pred_svc = svc_grid.predict(x_test)
precision_score(y_test, y_pred_svc, average="micro"), svc_grid


AdaBoost

# Hyperparameters
param = {"learning_rate": [0.0001, 0.001, 0.01, 0.1, 1.0, 2, 5],
       'n_estimators': [1,50,100, 500, 1000]}
#GridSearch

AdaBoost_Grid = GridSearchCV(AdaBoostClassifier(), param, cv= KFold(n_splits=5, random_state=42, shuffle = True),scoring = "accuracy")
AdaBoost_Grid.fit(x_train, y_train)
y_pred_ada = AdaBoost_Grid.predict(x_test)
precision_score(y_test, y_pred_ada, average="micro"), AdaBoost_Grid


Gradient Boost

# Hyperparameters
param = {"learning_rate": [0.01,0.1,1, 2, 5],
        'n_estimators': [100, 500, 1000],
        'max_depth': range(1,5),
        "max_features": ["auto", "sqrt", "log2"]}
#GridSearch
GradientBoost_Grid = GridSearchCV(GradientBoostingClassifier(), param, cv= KFold(n_splits=5, random_state=42, shuffle = True),scoring = "accuracy")
GradientBoost_Grid.fit(x_train, y_train)
y_pred_grad = GradientBoost_Grid.predict(x_test)
precision_score(y_test, y_pred_grad, average="micro"), GradientBoost_Grid


XG Boost

param = {'max_depth': [2, 6, 8], 'learning_rate': [0.01, 0.1],\
        'min_child_weight': [5, 6, 7], 'objective': ["multi:softprob"],\
        'subsample': [0.6, 0.4], "colsample_bytree": [0.7, 0.3],\
        "gamma": [0.5, 0.8, 1, 1.5], "n_estimators": [50, 100, 150],"eval_metric":["merror"]}

XgBoost_Grid = GridSearchCV(xgb.XGBClassifier(use_label_encoder=True, nthread=-1), param, cv= KFold(n_splits=5, random_state=42, shuffle = True), scoring ="accuracy")
XgBoost_Grid.fit(x_train, y_train)
y_pred_xgboost = XgBoost_Grid.predict(x_test)
precision_score(y_test, y_pred_xgboost, average="micro"), XgBoost_Grid


# Hyperparameters
param = {"max_depth": [2,6,8,10] ,\
        "min_samples_leaf": [1,2,5,7],\
        "n_estimators": [500,1500,3000],\
        "max_features":["auto", "sqrt","log2"]}
#GridSearch
grid_search = GridSearchCV(RandomForestClassifier(), param, n_jobs=-1, cv=5, scoring = "accuracy")
grid_search.fit(x_train, y_train)
y_pred_rfc = grid_search.predict(x_test)
precision_score(y_test, y_pred_rfc, average="micro"), grid_search


print("LDA : Best parameters {} with best score {} ".format(lda_grid.best_params_,\
lda_grid.best_score_))
print("QDA : Best parameters {} with best score {} ".format(qda_grid.best_params_,\
qda_grid.best_score_))
print("KNN : Best parameters {} with best score {} ".format(knn_grid.best_params_,\
knn_grid.best_score_))
print("SVC : Best parameters {} with best score {} ".format(svc_grid.best_params_,\
svc_grid.best_score_))
print("Random Forest : Best parameters {} with best score {} ".format(grid_search.best_params_,\
grid_search.best_score_))
print("XGBoost : Best parameters {} with best score {} ".format(XgBoost_Grid.best_params_,\
XgBoost_Grid.best_score_))
print("Gradient Boost : Best parameters {} with best score {} ".format(GradientBoost_Grid.best_params_,\
GradientBoost_Grid.best_score_))
print("ADA Boost : Best parameters {} with best score {} ".format(AdaBoost_Grid.best_params_,\
AdaBoost_Grid.best_score_))

import pickle
model_filename = "model_xg_grid.pkl"
with open(model_filename, 'wb') as file:
    pickle.dump(XgBoost_Grid, file)

model_filename = "../models/model_svc.pkl"
with open(model_filename, 'wb') as file:
    pickle.dump(svc, file)

model_filename = "../models/model_xg_grid.pkl"
with open(model_filename, 'wb') as file:
    pickle.dump(XgBoost_Grid, file)
